requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

torch==2.7.1+cu118
torchvision==0.22.1+cu118

pandas
matplotlib
transformers
pillow
scikit-learn
tensorboard
faiss-cpu
gradio